telegram_backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
psycopg==3.1.18
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==11.0.0
djangorestframework-simplejwt==5.3.1
django-filter==24.2
python-magic==0.4.27
django-storages==1.14.3
boto3==1.34.69
firebase-admin==6.5.0
